Update Navbar to handle Hugo Children
This enables the header's Navbar to take advantage of the Hugo `config.*` files' `menu.menu` child/parent relations. I added a simple static carat for the dropdowns as well since this is using bootstrap 3. As
I understand, bootstrap 4 adds the carat automatically, not really sure though.

The appearance of the dropdown menu isn't great, but I don't know CSS really, so I'm not sure how to clean this up. Feel free to make suggestions or edit yourself if you want.

This may fall outside of the scope of what this repo's vision is for minimal, but I wanted to propose it anyways so it ha the option to use more Hugo features.